Tests restruct
